Move all function inside a related folder
