+++ src/utils/GuidedFormFunctions/checkEmptyValues.ts
