+++ src/utils/GuidedFormFunctions/countChars.ts
