+++ src/utils/GuidedFormFunctions/countOneWord.ts
